requirements.txt
langchain
langchain-community
langchain-openai
streamlit
pypdf
faiss-cpu
python-dotenv
tiktoken